[refactor] config package
