+++ src/main/java/flab/project/config/RedisConfig.java
@@ -1,7 +1,7 @@
 package flab.project.config;
 
-import flab.project.data.dto.model.BasePost;
-import flab.project.data.dto.model.Profile;
+import flab.project.domain.post.model.BasePost;
+import flab.project.domain.user.model.Profile;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
